Organized tests and add a failing test for network
